Added private interface to libraries for improvement of testing
