+++ lib/CMakeLists.txt
@@ -1,0 +1,5 @@
+include_directories(src)
+
+add_subdirectory(src)
+add_subdirectory(tst)
+
