+++ lib/src/motion/motion.c
