+++ lib/tst/test_libmotion.c
